requirements.txt
streamlit==1.35.0
opencv-python-headless
mediapipe==0.10.9
python-dotenv
replicate
Pillow
requests
